"LNX.LW.2.0-00300-8x26.0"

Conflicts:
	AndroidKernel.mk
	Documentation/devicetree/bindings/pil/pil-q6v5-mss.txt
	arch/arm/boot/Makefile
	arch/arm/boot/dts/qcom/apq8026-v1-w-cdp.dts
	arch/arm/boot/dts/qcom/apq8026-v1-w-mtp.dts
	arch/arm/boot/dts/qcom/apq8074-dragonboard.dtsi
	arch/arm/boot/dts/qcom/apq8094.dtsi
	arch/arm/boot/dts/qcom/apqsamarium-mtp.dts
	arch/arm/boot/dts/qcom/dsi-panel-nt35590-qvga-cmd.dtsi
	arch/arm/boot/dts/qcom/fsm9900-cdp.dts
	arch/arm/boot/dts/qcom/fsm9900-mtp.dtsi
	arch/arm/boot/dts/qcom/fsm9900.dtsi
	arch/arm/boot/dts/qcom/mdm9630-v2-cdp.dts
	arch/arm/boot/dts/qcom/mpq8092-iommu.dtsi
	arch/arm/boot/dts/qcom/msm8226-memory.dtsi
	arch/arm/boot/dts/qcom/msm8226-v1-w-cdp.dts
	arch/arm/boot/dts/qcom/msm8226-v1-w-mtp.dts
	arch/arm/boot/dts/qcom/msm8226-v2-1080p-w-cdp.dts
	arch/arm/boot/dts/qcom/msm8226-v2-1080p-w-mtp.dts
	arch/arm/boot/dts/qcom/msm8226-v2-320p-w-cdp.dts
	arch/arm/boot/dts/qcom/msm8226-v2-320p-w-mtp.dts
	arch/arm/boot/dts/qcom/msm8226-w-memory.dtsi
	arch/arm/boot/dts/qcom/msm8226.dtsi
	arch/arm/boot/dts/qcom/msm8974.dtsi
	arch/arm/boot/dts/qcom/msm8994-sim.dts
	arch/arm/boot/dts/qcom/msmsamarium-cdp.dts
	arch/arm/boot/dts/qcom/msmsamarium-iommu.dtsi
	arch/arm/boot/dts/qcom/msmsamarium-mdss.dtsi
	arch/arm/boot/dts/qcom/msmsamarium-mtp.dts
	arch/arm/boot/dts/qcom/msmsamarium-rumi.dts
	arch/arm/boot/dts/qcom/msmsamarium-sim.dts
	arch/arm/boot/dts/qcom/msmsamarium.dtsi
	arch/arm/boot/dts/qcom/msmzirc-cdp.dts
	arch/arm/boot/dts/qcom/msmzirc-mtp.dts
	arch/arm/mach-msm/Makefile
	arch/arm/mm/dma-mapping.c
	arch/arm64/mm/dma-mapping.c
	drivers/char/adsprpc_compat.h
	drivers/char/hw_random/msm_rng.c
	drivers/gpu/msm/adreno-gpulist.h
	drivers/gpu/msm/adreno.c
	drivers/gpu/msm/adreno.h
	drivers/gpu/msm/adreno_a3xx.c
	drivers/gpu/msm/adreno_dispatch.c
	drivers/gpu/msm/adreno_drawctxt.c
	drivers/gpu/msm/adreno_profile.c
	drivers/gpu/msm/adreno_profile.h
	drivers/gpu/msm/adreno_ringbuffer.c
	drivers/gpu/msm/kgsl.c
	drivers/gpu/msm/kgsl_cffdump.c
	drivers/gpu/msm/kgsl_device.h
	drivers/gpu/msm/kgsl_events.c
	drivers/gpu/msm/kgsl_iommu.c
	drivers/gpu/msm/kgsl_pwrctrl.c
	drivers/gpu/msm/kgsl_snapshot.c
	drivers/input/misc/Kconfig
	drivers/input/touchscreen/synaptics_i2c_rmi4.c
	drivers/input/touchscreen/synaptics_i2c_rmi4.h
	drivers/platform/msm/qpnp-power-on.c
	drivers/power/Kconfig
	drivers/power/qpnp-charger.c
	drivers/power/reset/msm-poweroff.c
	drivers/soc/qcom/pil-msa.c
	drivers/soc/qcom/subsys-pil-tz.c
	drivers/soc/qcom/subsystem_restart.c
	drivers/soc/qcom/watchdog_v2.c
	drivers/tty/serial/msm_serial_hs.c
	drivers/usb/gadget/android.c
	drivers/usb/phy/phy-msm-usb.c
	drivers/video/msm/mdss/mdss_dsi.c
	drivers/video/msm/mdss/mdss_dsi.h
	drivers/video/msm/mdss/mdss_dsi_host.c
	drivers/video/msm/mdss/mdss_dsi_panel.c
	drivers/video/msm/mdss/mdss_fb.c
	drivers/video/msm/mdss/mdss_mdp.c
	drivers/video/msm/mdss/mdss_mdp.h
	drivers/video/msm/mdss/mdss_mdp_ctl.c
	drivers/video/msm/mdss/mdss_mdp_intf_cmd.c
	drivers/video/msm/mdss/mdss_mdp_overlay.c
	drivers/video/msm/mdss/mdss_mdp_pipe.c
	drivers/video/msm/mdss/mdss_panel.h
	drivers/video/msm/mdss/msm_mdss_io_8974.c
	include/linux/input/synaptics_dsx.h
	include/linux/sched.h
	include/linux/usb/msm_hsusb.h
	include/net/ping.h
	include/net/tcp.h
	include/soc/qcom/devfreq_devbw.h
	kernel/futex.c
	kernel/sched/core.c
	kernel/sched/sched.h
	kernel/sysctl.c
	net/ipv4/ping.c
	net/ipv4/raw.c
	net/ipv4/sysctl_net_ipv4.c
	net/ipv6/ping.c
	net/ipv6/route.c
	net/wireless/nl80211.c
	sound/soc/codecs/wcd9306.c
	sound/soc/msm/msm8226.c
	sound/soc/msm/qdsp6v2/msm-dai-q6-v2.c

+++ include/trace/events/ice40.h
@@ -254,15 +254,15 @@ TRACE_EVENT(ice40_setup,
 
 TRACE_EVENT(ice40_in,
 
-	TP_PROTO(u16 ep, const char *token, u8 len, u8 expected, int ret),
+	TP_PROTO(u16 ep, const char *token, u32 len, u32 expected, int ret),
 
 	TP_ARGS(ep, token, len, expected, ret),
 
 	TP_STRUCT__entry(
 		__field(u16, ep)
 		__string(token, token)
-		__field(u8, len)
-		__field(u8, expected)
+		__field(u32, len)
+		__field(u32, expected)
 		__field(int, ret)
 	),
 
@@ -282,14 +282,14 @@ TRACE_EVENT(ice40_in,
 
 TRACE_EVENT(ice40_out,
 
-	TP_PROTO(u16 ep, const char *token, u8 len, int ret),
+	TP_PROTO(u16 ep, const char *token, u32 len, int ret),
 
 	TP_ARGS(ep, token, len, ret),
 
 	TP_STRUCT__entry(
 		__field(u16, ep)
 		__string(token, token)
-		__field(u8, len)
+		__field(u32, len)
 		__field(int, ret)
 	),
